Don't run cronjobs on DEV

+++ django/docker-entrypoint.sh
@@ -26,17 +26,19 @@ waitfordb()
 
 CMD="$1"
 
-waitfordb
->&2 echo "Starting crond in the background"
-# Add our cronjob
-crontab -l > mycron
-cat /app/demsausage/app/sausage/cron/crontab.txt >> mycron
-crontab mycron
-rm mycron
+if [ "$ENVIRONMENT" != "DEVELOPMENT" ]; then
+  waitfordb
+  >&2 echo "Starting crond in the background"
+  # Add our cronjob
+  crontab -l > mycron
+  cat /app/demsausage/app/sausage/cron/crontab.txt >> mycron
+  crontab mycron
+  rm mycron
 
-# Start crond
-chmod 755 /app/demsausage/app/sausage/cron/cron.sh
-crond
+  # Start crond
+  chmod 755 /app/demsausage/app/sausage/cron/cron.sh
+  crond
+fi
 
 if [ "$ENVIRONMENT" = "DEVELOPMENT" ]; then
   waitfordb
